CMSSW Resources Utilisation v2
Remove the make_circles.py script, and refactor its functionality into the
FastTimerService and the web page itself.

The FastTimerService now (since CMSSW_11_1_0_pre5) produces an output in JSON
format,suitable for uploading to the web page.

The web page chooses the metric to display, creates the module hierarchiy and
computes the partial sums, and applies the colour scheme on the fly.

The advantages are:
  - no need to have a full dump of the configuration to extract the module types;
  - no need to run an additinal script to produce the hierarchy;
  - on the fly choice of metrics, grouping and colour scheme.