Implement barchart view (#42)
Implement a bar chart view as an alternative to the default pie view to visualise
CMSSW resources utilisation. The grouping and colors logic have remained the same
so that changes to either should be transparent for both visualisations.

By default, when entering the bar view (selectable via a new drop-down menu) two
plots will be visible: one with a bar view of all groups and one with the same
information collapsed in a single stacked plot.

On the right side of the page, this view shows the same table found on the regular
pie chart view, allowing to change the order of both plots.

One more option available in this view is found below the table, where other
datasets can be selected to be compared with the currently loaded one. After at
least another dataset is selected, its stacked plot is shown in the bottom plot
and the diff view becomes available. Selecting this option changes the top and
bottom plot to show side-by-side stacks for each group in the selected dataset
and the corresponding difference, respectively.

While in diff view, the difference can be toggled between absolute and relative (%).
Up to 10 different datasets can be compared at once. Moreover, focusing on a
single group, the top 15 modules by resource utilisation are compared.

Minimal changes have been applied to the existing files to support the switch
between pie and chart view.